fix pr comments: remove parallelize, at most once impl
