+++ quickwit/quickwit-indexing/src/source/gcp_pubsub_source.rs
@@ -17,35 +17,31 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+use std::borrow::BorrowMut;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use anyhow::{anyhow, Context};
 use async_trait::async_trait;
-use bytes::Bytes;
 use google_cloud_default::WithAuthExt;
 use google_cloud_pubsub::client::{Client, ClientConfig};
-use google_cloud_pubsub::subscriber::ReceivedMessage;
 use google_cloud_pubsub::subscription::Subscription;
 use quickwit_actors::{ActorContext, ActorExitStatus, Mailbox};
 use quickwit_config::GcpPubSubSourceParams;
-use quickwit_metastore::checkpoint::{
-    PartitionId, Position, SourceCheckpoint, SourceCheckpointDelta,
-};
+use quickwit_metastore::checkpoint::{PartitionId, Position, SourceCheckpoint};
 use serde_json::Value as JsonValue;
-use tokio::sync::RwLock;
 use tokio::time;
 use tracing::{info, warn};
-use uuid::Uuid;
 
 use super::SourceActor;
 use crate::actors::DocProcessor;
-use crate::models::{NewPublishLock, PublishLock, RawDocBatch};
-use crate::source::{Source, SourceContext, SourceExecutionContext, TypedSourceFactory};
+use crate::source::{
+    BatchBuilder, Source, SourceContext, SourceExecutionContext, TypedSourceFactory,
+};
 
 const BATCH_NUM_BYTES_LIMIT: u64 = 5_000_000;
 const DEFAULT_MAX_MESSAGES_PER_PULL: i32 = 1000;
-const DEFAULT_PULL_PARALLELISM: u64 = 10; // TODO: is 10 too high as a default?
+// const DEFAULT_PULL_PARALLELISM: u64 = 10; // TODO: is 10 too high as a default?
 
 pub struct GcpPubSubSourceFactory;
 
@@ -65,9 +61,8 @@ impl TypedSourceFactory for GcpPubSubSourceFactory {
 
 #[derive(Default)]
 pub struct GcpPubSubSourceState {
-    ack_ids: RwLock<Vec<String>>,
     subscription_is_empty: bool,
-    doc_count: RwLock<u64>,
+    doc_count: u64,
 }
 
 pub struct GcpPubSubSource {
@@ -76,9 +71,9 @@ pub struct GcpPubSubSource {
     state: GcpPubSubSourceState,
     subscription_source: Subscription,
     backfill_mode_enabled: bool,
-    publish_lock: PublishLock,
     partition_id: String,
-    pull_parallelism: u64,
+    // Parallelism will be enabled in next PR
+    // pull_parallelism: u64,
     max_messages_per_pull: i32,
 }
 
@@ -89,10 +84,7 @@ impl GcpPubSubSource {
     ) -> anyhow::Result<Self> {
         let subscription = params.subscription.clone();
         let backfill_mode_enabled = params.enable_backfill_mode;
-        let pull_parallelism = match params.pull_parallelism {
-            Some(parallelism) => parallelism,
-            None => DEFAULT_PULL_PARALLELISM,
-        };
+        // let pull_parallelism = params.pull_parallelism.unwrap_or(DEFAULT_PULL_PARALLELISM);
         let max_messages_per_pull = match params.max_messages_per_pull {
             Some(max) => max,
             None => DEFAULT_MAX_MESSAGES_PER_PULL,
@@ -109,13 +101,12 @@ impl GcpPubSubSource {
             .await
             .expect("Failed to create GcpPubSub client");
         let subscription_source = client.subscription(&subscription);
-        let publish_lock = PublishLock::default();
 
         info!(
             index_id=%ctx.index_uid.index_id(),
             source_id=%ctx.source_config.source_id,
             subscription=%subscription,
-            parallelism=%pull_parallelism,
+            // parallelism=%pull_parallelism,
             "Starting GcpPubSub source."
         );
 
@@ -125,9 +116,9 @@ impl GcpPubSubSource {
             state: GcpPubSubSourceState::default(),
             subscription_source,
             backfill_mode_enabled,
-            publish_lock,
-            partition_id: Uuid::new_v4().to_string(),
-            pull_parallelism,
+            // TODO: get the real value
+            partition_id: "<node_id>/<pipeline_ord>".to_string(),
+            // pull_parallelism,
             max_messages_per_pull,
         })
     }
@@ -137,58 +128,8 @@ impl GcpPubSubSource {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-struct BatchBuilder {
-    docs: Vec<Bytes>,
-    num_bytes: u64,
-    checkpoint_delta: SourceCheckpointDelta,
-}
-
-impl BatchBuilder {
-    fn build(self) -> RawDocBatch {
-        RawDocBatch {
-            docs: self.docs,
-            checkpoint_delta: self.checkpoint_delta,
-            force_commit: false,
-        }
-    }
-
-    fn build_force(self) -> RawDocBatch {
-        RawDocBatch {
-            docs: self.docs,
-            checkpoint_delta: self.checkpoint_delta,
-            force_commit: true,
-        }
-    }
-
-    // fn clear(&mut self) {
-    //     self.docs.clear();
-    //     self.num_bytes = 0;
-    //     self.checkpoint_delta = SourceCheckpointDelta::default();
-    // }
-
-    fn push(&mut self, message: ReceivedMessage) {
-        let doc = message.message.data;
-        self.num_bytes += doc.len() as u64;
-        self.docs.push(doc.into());
-    }
-}
-
 #[async_trait]
 impl Source for GcpPubSubSource {
-    async fn initialize(
-        &mut self,
-        doc_processor_mailbox: &Mailbox<DocProcessor>,
-        ctx: &SourceContext,
-    ) -> Result<(), ActorExitStatus> {
-        info!("GcpPubSub initializing");
-        let publish_lock = self.publish_lock.clone();
-        ctx.send_message(doc_processor_mailbox, NewPublishLock(publish_lock))
-            .await?;
-        info!("GcpPubSub initialized");
-        Ok(())
-    }
-
     async fn emit_batches(
         &mut self,
         doc_processor_mailbox: &Mailbox<DocProcessor>,
@@ -196,22 +137,16 @@ impl Source for GcpPubSubSource {
     ) -> Result<Duration, ActorExitStatus> {
         info!("GcpPubSub beginning batch");
         let now = Instant::now();
-        let batch_lock = Arc::new(RwLock::new(BatchBuilder::default()));
+        let mut batch: BatchBuilder = BatchBuilder::default();
         let deadline = time::sleep(*quickwit_actors::HEARTBEAT / 2);
         tokio::pin!(deadline);
 
-        info!("GcpPubSub pulling batch");
-
+        // TODO: lets add parallelism support in next PR
+        // TODO: ensure we ACK the message after being commit: at least once
+        // TODO: ensure we increase_ack_deadline for the items
         loop {
-            let mut handles = vec![];
-
-            for _ in 1..self.pull_parallelism {
-                handles.push(self.pull_message_batch(Arc::clone(&batch_lock)));
-            }
-
             tokio::select! {
-                _ = futures::future::join_all(handles) => {
-                    let batch = batch_lock.read().await;
+                _ = self.pull_message_batch(batch.borrow_mut()) => {
                     if batch.num_bytes >= BATCH_NUM_BYTES_LIMIT {
                         break;
                     }
@@ -223,13 +158,11 @@ impl Source for GcpPubSubSource {
             ctx.record_progress();
         }
 
-        let batch = batch_lock.read().await.clone(); // TODO: This clone is wasteful! There must be a better way
         if self.should_exit() {
             info!(subscription = %self.subscription, "Reached end of subscription.");
             ctx.ask(doc_processor_mailbox, batch.build_force())
                 .await
                 .context("Failed to force commit last batch!")?;
-            self.ack_ids().await?;
             ctx.send_exit_with_success(doc_processor_mailbox).await?;
             return Err(ActorExitStatus::Success);
         }
@@ -254,16 +187,8 @@ impl Source for GcpPubSubSource {
         _checkpoint: SourceCheckpoint,
         _ctx: &ActorContext<SourceActor>,
     ) -> anyhow::Result<()> {
-        // TODO: How can we know if these are ok to ack? We need to use the checkpoint!
-        self.ack_ids().await
-    }
-
-    async fn finalize(
-        &mut self,
-        _exit_status: &ActorExitStatus,
-        _ctx: &SourceContext,
-    ) -> anyhow::Result<()> {
-        self.ack_ids().await
+        // TODO: add ack of ids
+        anyhow::Ok(())
     }
 
     fn name(&self) -> String {
@@ -279,33 +204,7 @@ impl Source for GcpPubSubSource {
 }
 
 impl GcpPubSubSource {
-    async fn ack_ids(&mut self) -> anyhow::Result<()> {
-        let mut ack_ids = self.state.ack_ids.write().await;
-        if ack_ids.is_empty() {
-            return Ok(());
-        }
-
-        // TODO: For ordered GcpPubSub topics/subscriptions we need to ensure we
-        // also ack in that same order!
-        // TODO: We may have consumed more messages by the time we get a reply!
-        // We can't just blindly clear... instead we need to keep track of which
-        // ack ids were present in each checkpoint
-        info!("Acking ids...");
-        for chunk in ack_ids.chunks(1000) {
-            self.subscription_source
-                .ack(Vec::from(chunk))
-                .await
-                .map_err(anyhow::Error::from)?;
-        }
-        ack_ids.clear();
-        info!("Acked!");
-        Ok(())
-    }
-
-    async fn pull_message_batch(
-        &self,
-        batch_lock: Arc<RwLock<BatchBuilder>>,
-    ) -> anyhow::Result<()> {
+    async fn pull_message_batch(&mut self, batch: &mut BatchBuilder) -> anyhow::Result<()> {
         let event = self
             .subscription_source
             .pull(self.max_messages_per_pull, None)
@@ -318,19 +217,17 @@ impl GcpPubSubSource {
 
         let length = messages.len();
         info!("GcpPubSub pulled {length} messages");
-        let mut batch = batch_lock.write().await;
-        let mut ack_ids = self.state.ack_ids.write().await;
-        let mut doc_count = self.state.doc_count.write().await;
-        let initial_position = *doc_count;
+        let initial_position = self.state.doc_count;
 
         for message in messages {
-            ack_ids.push(String::from(message.ack_id()));
-            batch.push(message);
-            *doc_count += 1
+            message.ack().await?; // TODO: remove ACK here when doing at least once
+            let num_bytes = message.message.data.len() as u64;
+            batch.push(message.message.data.into(), num_bytes);
+            self.state.doc_count += 1
         }
 
         let first_position = Position::from(initial_position);
-        let last_position = Position::from(*doc_count);
+        let last_position = Position::from(self.state.doc_count);
         let partition_id = PartitionId::from(self.partition_id.clone());
 
         if first_position != last_position {
@@ -339,17 +236,6 @@ impl GcpPubSubSource {
                 .record_partition_delta(partition_id, first_position, last_position)
                 .context("Failed to record partition delta.")?;
         }
-
         Ok(())
     }
-
-    // async fn increase_ack_deadline(self) {
-    //     todo!("google-cloud-GcpPubSub doesn't implement this...")
-    //     // BUT it does!... kind of...
-    //     // It has the ability to do this for a subscriber (streaming pull).
-    //     // We could potentially make a hacky copy of how it does that...
-    //     // Or PR in this functionality into the Subscription struct properly.
-    //     //
-    //     // For now we can just set the deadline to 600 seconds on gcp and call it day?
-    // }
 }
